==> 0. File locations.R <==
#MHSDS
rawdatadir <- "M:/Analytics/CYP MH/England/MHSDS/" #Replace with name of your MHSDS data folder and start with an EMPTY folder
main_name <- "Main performance files" #Change this if you would like your sub-folder to have a different name
ed_name <- "Eating disorders files" #Change this if you would like your sub-folder to have a different name

#NHS workforce
workforcedir <- "M:/Analytics/Workforce statistics/"
vacanciesdir <- "M:/Analytics/NHS vacancy statistics/"

#Scotland
rawdatadir_scot <- "M:/Analytics/CYP MH/Scotland/"

#Wales
rawdatadir_wales <- "M:/Analytics/CYP MH/Wales/"
